# docs/swagger.yaml
openapi: 3.0.0
servers:
  - url: "https://api.eesast.com"
    description: Remote API Server
  - url: "http://localhost:{port}"
    description: Local API Server
    variables:
      port:
        description: 端口号
        default: "28888"
info:
  title: API Backend for EESAST
  description: |
    清华大学电子工程系学生科协后端 API。了解更多 EESAST 内容，请访问：[GitHub 主页](https://github.com/eesast/)或 [EESAST 官方网站](https://eesast.com/)。
  version: 1.0.0
  contact:
    name: API Support
    email: i@raineggplant.com
...
paths:
...
  "/v1/rooms":
    get:
      tags:
        - rooms
      summary: 返回匹配查询的房间
      description: 可以通过 query 进行筛选
      operationId: findRooms
      parameters:
        - name: contestId
          in: query
          schema:
            type: number
          description: 比赛ID
        - name: status
          in: query
          schema:
            type: number
            enum: [0, 1, 2]
          description: 房间是否开启
        - name: begin
          in: query
          schema:
            type: number
          description: 按时间降序排列，起始房间的位置
        - name: end
          in: query
          schema:
            type: number
          description: 按时间降序排列，结束房间的位置
      responses:
        200:
          description: Successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Room"
        400:
          $ref: "#/components/responses/BadRequest"
        500:
          description: Server error

    post:
      tags:
        - rooms
      summary: 开启一个新房间
      description: 需要已登录用户
      operationId: addRoom
      security:
        - bearerAuth: []
      requestBody:
        $ref: "#/components/requestBodies/addRoom"
      responses:
        201:
          description: Successful operation
          headers:
            Location:
              description: 新开启房间的 URL
              schema:
                type: string
        400:
          # TODO: Solve status code conflict
          # $ref: "#/components/responses/BadRequest"
          description: "TODO: Solve status code conflict. --- Contest not available"
        401:
          $ref: "#/components/responses/UnauthorizedError"
        500:
          description: Server error
        503:
          description: Failed to start docker container
  "/v1/rooms/{roomId}":
    get:
      tags:
        - rooms
      summary: 根据 ID 获取房间
      operationId: getRoomById
      parameters:
        - name: roomId
          in: path
          description: 房间 ID
          required: true
          schema:
            type: number
      responses:
        200:
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Room"
        400:
          $ref: "#/components/responses/BadRequest"
        500:
          description: Server error

    put:
      tags:
        - rooms
      summary: 更新已存在的房间
      description: 需要 `root` 或 `organizer` 角色
      operationId: updateRoom
      security:
        - bearerAuth: []
      parameters:
        - name: roomId
          in: path
          description: 待更新房间的 ID
          required: true
          schema:
            type: number
      requestBody:
        $ref: "#/components/requestBodies/updateRoom"
      responses:
        204:
          description: Successful operation
          headers:
            Location:
              description: 被更新比赛的 URL
              schema:
                type: string
        401:
          $ref: "#/components/responses/UnauthorizedError"
        404:
          description: "Not Found: Room does not exist"
        500:
          description: Server error

    delete:
      tags:
        - rooms
      summary: 删除一个房间
      description: 需要 `root` 或 `organizer` 角色
      operationId: deleteRoom
      security:
        - bearerAuth: []
      parameters:
        - name: roomId
          in: path
          description: 要删除房间的 ID
          required: true
          schema:
            type: number
      responses:
        204:
          description: Successful operation
        401:
          $ref: "#/components/responses/UnauthorizedError"
        404:
          description: "Not Found: Room does not exist / Docker container does not exist"
        422:
          description: "Unprocessable Entity: Missing form data"
        500:
          description: Server error
        503:
          description: "Service Unavailable: Failed to stop or remove docker container"
  "/v1/rooms/check/{token}":
    post:
      tags:
        - rooms
      summary: 验证服务器启动的token
      operationId: checkRoomToken
      parameters:
        - name: token
          in: path
          required: true
          schema:
            type: string
      responses:
        204:
          description: Successful operation
        400:
          $ref: "#/components/responses/BadRequest"
        401:
          description: Wrong token
        500:
          description: Server error
  "/v1/rooms/{roomId}/join":
    post:
      tags:
        - rooms
      summary: 加入房间
      description: 由server发起，返回更新后队伍数组，header中为server token
      operationId: joinRoom
      security:
        - bearerAuth: []
      parameters:
        - name: roomId
          in: path
          description: 加入房间的 ID
          required: true
          schema:
            type: number
      requestBody:
        $ref: "#/components/requestBodies/changeRoomTeam"
      responses:
        200:
          description: Successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  type: number
        400:
          # TODO: Solve status code conflict
          # $ref: "#/components/responses/BadRequest"
          description: "TODO: Solve status code conflict. --- User not in team"
        401:
          $ref: "#/components/responses/UnauthorizedError"
        403:
          description: "Forbidden: Permission denied"
        404:
          description: "Not Found: Room does not exist"
        409:
          description: Team already in room
        500:
          description: Server error
  "/v1/rooms/{roomId}/leave":
    post:
      tags:
        - rooms
      summary: 退出房间
      description: 由server发起，返回剩余队伍数组，header中为server token
      operationId: leaveRoom
      security:
        - bearerAuth: []
      parameters:
        - name: roomId
          in: path
          description: 退出房间的 ID
          required: true
          schema:
            type: number
      requestBody:
        $ref: "#/components/requestBodies/changeRoomTeam"
      responses:
        200:
          description: Successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  type: number
        400:
          # TODO: Solve status code conflict
          # $ref: "#/components/responses/BadRequest"
          description: "TODO: Solve status code conflict. --- User not in team"
        401:
          $ref: "#/components/responses/UnauthorizedError"
        403:
          description: "Forbidden: Permission denied"
        404:
          description: "Not Found: Room does not exist"
        500:
          description: Server error
  "/v1/rooms/{roomId}/status":
    put:
      tags:
        - rooms
      summary: 更新房间状态
      description: 由server发起
      operationId: updateRoomStatus
      security:
        - bearerAuth: []
      parameters:
        - name: roomId
          in: path
          description: 房间 ID
          required: true
          schema:
            type: number
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - status
              properties:
                status:
                  type: number
                  enum: [0, 1, 2]
      responses:
        401:
          $ref: "#/components/responses/UnauthorizedError"
        403:
          description: "Forbidden: Permission denied"
        404:
          description: "Not Found: Room does not exist / Docker container does not exist"
        422:
          description: "Unprocessable Entity: Missing form data"
        500:
          description: Server error
        503:
          description: "Service Unavailable: Failed to stop or remove docker container"
...
components:
  schemas:
...
    Date:
      type: string
      format: date-time
      description: >-
        日期对象。使用时采用 ISO
        格式字符串表示。[更多信息](https://www.iso.org/iso-8601-date-and-time-format.html)
...
    Room:
      type: object
      required:
        - id
        - contestId
        - teams
        - ip
        - port
      properties:
        id:
          type: number
          description: 自增值
        contestId:
          type: number
          description: 房间ID
        status:
          type: number
          enum: [0, 1, 2]
          default: 0
          description: 房间状态，0表示未开启，1表示已开启，2表示已关闭
        teams:
          type: array
          items:
            type: number
            description: 队伍ID(机器人可用-1表示)
        ip:
          type: string
          description: ip地址
        port:
          type: number
          description: 端口
        createdAt:
          $ref: "#/components/schemas/Date"
        createdBy:
          type: number
        updatedAt:
          $ref: "#/components/schemas/Date"
        updatedBy:
          type: number
...
  requestBodies:
...
    addRoom:
      content:
        application/json:
          schema:
            type: object
            required:
              - contestId
              - teams
              - ip
              - port
            properties:
              contestId:
                type: number
                description: 比赛ID
              teams:
                type: array
                items:
                  type: number
                  description: 队伍ID(机器人可用-1表示)
              ip:
                type: string
                description: ip地址
              port:
                type: number
                description: 端口
      description: 新建房间的提交格式
      required: true
    updateRoom:
      content:
        application/json:
          schema:
            type: object
            properties:
              contestId:
                type: number
                description: 比赛ID
              teams:
                type: array
                items:
                  type: number
                  description: 队伍ID(机器人可用-1表示)
              ip:
                type: string
                description: ip地址
              port:
                type: number
                description: 端口
      description: 更新房间的提交格式
      required: true
    changeRoomTeam:
      content:
        application/json:
          schema:
            type: object
            required:
              - token
            properties:
              token:
                type: string
                description: 用户token
      description: 更改房间中队伍的提交格式
      required: true
...
  responses:
    BadRequest:
      content:
        application/json:
          schema:
            type: object
            properties:
              errors:
                type: array
                description: 错误列表
                items:
                  type: object
                  properties:
                    path:
                      type: string
                      description: 错误位置
                      example: ".params.articleId"
                    message:
                      type: string
                      description: 错误信息
                      example: "should be number"
                    errorCode:
                      type: string
                      description: 错误码
                      example: "type.openapi.validation"
      description: 包含请求错误信息的 JSON
    UnauthorizedError:
      description: "Access token is missing, invalid or expired / Permission denied"
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT